userInterface/requirements.txt
streamlit>=1.31.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-ollama>=0.1.0
langchain-text-splitters>=0.0.1
chromadb>=0.4.0
pypdf2>=3.0.0
python-docx>=0.8.11
pandas>=1.5.0
python-multipart>=0.0.5
requests>=2.28.1
beautifulsoup4>=4.11.1
python-dotenv>=1.0.0
ollama>=0.1.0
